do same to mesh
